chore: simplify file-size check to a flat 1000-line limit
Replace the homegrown per-file override sprawl with a single shared
check. The 500-line limit had grown ~50 grandfathered overrides in
desktop, turning the escape hatch into the norm.

- Extract the walk/count/violation logic into a shared core at
  scripts/check-file-sizes-core.mjs; desktop and web now import it so
  the two checks can't drift. The core also guards missing root dirs
  (previously only web did).
- Raise the limit to a flat 1000 per rule. No warnings, no two-tier.
- Drop all overrides except the 7 desktop files that genuinely exceed
  1000; those stay as TEMP entries (at current line count) to be split
  in a follow-up. Web has no overrides.

pnpm -r check is green: biome clean + both file-size checks pass.
